Added reliable delivery to the codebase
Reliable delivery is implemented using the DeliveryNotificationManager
which reports on whether packets are delivered or not. If packets are
reported as undelivered it is up to the sender to decide what to do. You
can inherit from TransmissionData and override HandleDeliveryFailure()
and HandleDeliverySuccess() to decide what happens when messages are
dropped